Final cleanups in gradle build setup, worked through deprecated and unchecked warnings.
The gradle build was a bit noisy because of various limitations and/or bugs in Gradle
1.10, which is used in the Vert.x gradle build template. Additionally, the log4j setup
in netty (a dependency for netty) was complaining about a log4j Appender not being
available. These things are now fixed so we have a cleaner build experience.

The new Neo4J library (2.1) has deprecated the Transaction.finish() method in favour
of either using Transaction.close() explicitly or by using a try-with-resources
statement that automatically calls either Transaction.failure when the try fails
and always calls Transaction.close() regardless of Transaction.succes() having run.
I've opted to keep the code JDK6 compatible so I did not use the try-with-resources
method and instead added a few missing Transaction.succes() statements and changed
the Transaction.finish() statements to Transaction.close().